separate components using graphql api
